first version - splitting to debug files
