+++ chain/client-primitives/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod debug;
 pub mod types;
