+++ chain/client/src/lib.rs
@@ -1,12 +1,13 @@
 pub use near_client_primitives::types::{
-    DebugStatus, Error, GetBlock, GetBlockHash, GetBlockProof, GetBlockProofResponse,
-    GetBlockWithMerkleTree, GetChunk, GetExecutionOutcome, GetExecutionOutcomeResponse,
-    GetExecutionOutcomesForBlock, GetGasPrice, GetNetworkInfo, GetNextLightClientBlock,
-    GetProtocolConfig, GetReceipt, GetStateChanges, GetStateChangesInBlock,
-    GetStateChangesWithCauseInBlock, GetStateChangesWithCauseInBlockForTrackedShards,
-    GetValidatorInfo, GetValidatorOrdered, Query, QueryError, Status, StatusResponse, SyncStatus,
-    TxStatus, TxStatusError,
+    Error, GetBlock, GetBlockHash, GetBlockProof, GetBlockProofResponse, GetBlockWithMerkleTree,
+    GetChunk, GetExecutionOutcome, GetExecutionOutcomeResponse, GetExecutionOutcomesForBlock,
+    GetGasPrice, GetNetworkInfo, GetNextLightClientBlock, GetProtocolConfig, GetReceipt,
+    GetStateChanges, GetStateChangesInBlock, GetStateChangesWithCauseInBlock,
+    GetStateChangesWithCauseInBlockForTrackedShards, GetValidatorInfo, GetValidatorOrdered, Query,
+    QueryError, Status, StatusResponse, SyncStatus, TxStatus, TxStatusError,
 };
+
+pub use near_client_primitives::debug::DebugStatus;
 
 pub use crate::client::Client;
 pub use crate::client_actor::{start_client, ClientActor};
@@ -15,6 +16,7 @@ pub use crate::view_client::{start_view_client, ViewClientActor};
 pub mod adversarial;
 mod client;
 mod client_actor;
+pub mod debug;
 mod info;
 mod metrics;
 mod rocksdb_metrics;
