+++ chain/jsonrpc-primitives/src/types/status.rs
@@ -8,7 +8,7 @@ pub struct RpcStatusResponse {
 
 #[derive(Debug, Serialize)]
 pub struct RpcDebugStatusResponse {
-    pub status_response: near_client_primitives::types::DebugStatusResponse,
+    pub status_response: near_client_primitives::debug::DebugStatusResponse,
 }
 
 #[derive(Debug, Serialize, Deserialize)]
