+++ core/primitives/src/views.rs
@@ -323,32 +323,6 @@ pub struct ValidatorInfo {
 }
 
 #[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DebugChunkStatus {
-    pub shard_id: u64,
-    pub chunk_hash: ChunkHash,
-    pub chunk_producer: String,
-    pub gas_used: u64,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub processing_time_ms: Option<u64>,
-}
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DebugBlockStatus {
-    pub block_hash: CryptoHash,
-    pub block_height: u64,
-    pub block_producer: String,
-    pub chunks: Vec<DebugChunkStatus>,
-    // Time that was spent processing a given block.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub processing_time_ms: Option<u64>,
-    // Time between this block and the next one in chain.
-    pub timestamp_delta: u64,
-    pub gas_price_ratio: f64,
-}
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
 #[derive(Serialize, Deserialize, Debug, PartialEq, Eq)]
 pub struct PeerInfoView {
     pub addr: String,
@@ -399,18 +373,6 @@ impl From<Tip> for BlockStatusView {
 
 #[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
 #[derive(Serialize, Deserialize, Debug)]
-pub struct EpochInfoView {
-    pub epoch_id: CryptoHash,
-    pub height: BlockHeight,
-    pub first_block: Option<(CryptoHash, DateTime<chrono::Utc>)>,
-    pub validators: Vec<ValidatorInfo>,
-    pub chunk_only_producers: Vec<String>,
-    pub protocol_version: u32,
-    pub shards_size_and_parts: Vec<(u64, u64, bool)>,
-}
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Serialize, Deserialize, Debug)]
 pub struct BlockByChunksView {
     pub height: BlockHeight,
     pub hash: CryptoHash,
@@ -425,13 +387,6 @@ pub struct ChunkInfoView {
     pub num_of_chunks_in_progress: usize,
     pub num_of_orphans: usize,
     pub next_blocks_by_chunks: Vec<BlockByChunksView>,
-}
-
-#[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
-#[derive(Serialize, Deserialize, Debug)]
-pub struct TrackedShardsView {
-    pub shards_tracked_this_epoch: Vec<bool>,
-    pub shards_tracked_next_epoch: Vec<bool>,
 }
 
 #[cfg_attr(feature = "deepsize_feature", derive(deepsize::DeepSizeOf))]
